# validation: improve performance of CheckBlockIndex