@martinus 2023-08-25
Hi, I wonder how you measured the runtime, is there a benchmark? Another optimization that might be worth it, it is to replace the `[multimap](std::multimap<CBlockIndex*,CBlockIndex*> forward)` with an `std::unordered_map<CBlockIndex*, std::vector<CBlockIndex*>> forward`.