@Sjors 2023-09-14
While you're at it, I noticed that the help text for `-checkblockindex` says "occasionally" (added in #7290), but afaik it happens every block (and every header during the initial header sync). We should probably drop that word, or alternatively provide an `interval` argument.