@mzumsande 2023-09-15
> While you're at it, I noticed that the help text for -checkblockindex says "occasionally" (added in https://github.com/bitcoin/bitcoin/pull/7290), but afaik it happens every block (and every header during the initial header sync). We should probably drop that word, or alternatively provide an interval argument.

Good point, I think that an interval argument similar to addrman or mempool consistency checks makes sense. Will try that in addition to the other points above (it might take a few weeks though, therefore put the PR into draft).