@mzumsande 2023-12-22
Benchmark results:
<details>
<summary>master</summary>

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|          331,165.67 |            3,019.64 |    0.3% |      0.01 | `CheckBlockIndex`
|          320,603.00 |            3,119.12 |    1.7% |      0.01 | `CheckBlockIndex`
|          315,683.67 |            3,167.73 |    0.4% |      0.01 | `CheckBlockIndex`
|          329,077.33 |            3,038.80 |    0.3% |      0.01 | `CheckBlockIndex`
|          315,345.33 |            3,171.13 |    0.4% |      0.01 | `CheckBlockIndex`

</details>

<details>
<summary>PR</summary>

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|          127,273.00 |            7,857.13 |    1.8% |      0.01 | `CheckBlockIndex`
|          123,687.22 |            8,084.91 |    0.1% |      0.01 | `CheckBlockIndex`
|          136,637.17 |            7,318.65 |    0.1% |      0.01 | `CheckBlockIndex`
|          123,805.67 |            8,077.17 |    0.3% |      0.01 | `CheckBlockIndex`
|          124,162.62 |            8,053.95 |    0.2% |      0.01 | `CheckBlockIndex`
</details>

<details>
<summary>martinus' idea</summary>

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|          384,395.00 |            2,601.49 |    0.5% |      0.01 | `CheckBlockIndex`
|          380,674.67 |            2,626.92 |    0.3% |      0.01 | `CheckBlockIndex`
|          384,337.33 |            2,601.88 |    1.1% |      0.01 | `CheckBlockIndex`
|          410,836.50 |            2,434.06 |    3.4% |      0.01 | `CheckBlockIndex`
|          384,461.33 |            2,601.04 |    2.0% |      0.01 | `CheckBlockIndex`
</details>

The benchmark currently uses a linear chain of 1100 headers. In general, the efficiency gain becomes larger with a longer chain, for example with just 100 headers the speedup is less.

So there is a speedup of a facto >2.5 with the benchmark (that becomes larger on larger chains, see results for mainnet / testnet in OP).
The suggestion by @martinus to use `std::unordered_map<CBlockIndex*, std::vector<CBlockIndex*>>` doesn't seem to improve performance (though it's also very possible that I might have not implemented it in a great away).