@mzumsande 2024-03-20
[4f51a4c ](https://github.com/bitcoin/bitcoin/commit/4f51a4c01f2ce8ee1390e3b7a373d91158d8a83e)to [7fcd686](https://github.com/bitcoin/bitcoin/commit/7fcd686f1fe219d8c4cb4ba90127d1139eee9d22): Rebased, addrressed nit.

> Would be nice to cover this function properly with a good number of unit tests. It shouldn't be hard, the `validation_block_tests.cpp` file contain all the util functions to create different chain topologies.

I agree, might take me a few weeks to get to that thoughs since I have a few other things on my list.
Although just mentioning that quite a few constellations are already covered in random unit / functional tests that are targeted at something else - when I implemented this originally and didn't get some details right, I got several failures.