@mzumsande 2024-04-01
[b51eb42](https://github.com/bitcoin/bitcoin/commit/b51eb422d657bb1d0e6f9c7e71eb8a87a8c74f42) to [e880ee0](https://github.com/bitcoin/bitcoin/commit/e880ee0a634f0be5079d39328fa5e118452553a9)
Addressed feedback by @ryanofsky - now the diff to master should be a bit smaller than before.

Still haven't had the time to write more unit tests for `CheckBlockIndex()`.