@hebasto 2024-07-31
Ported to the CMake-based build system in https://github.com/hebasto/bitcoin/pull/290.